Fix the build by locking versions during the container build
This adds the lock file to the build container, to use the expected versions and
not let the build process select newer ones.

I believe this is ok, as the container build already selects a fixed build and runtime
container tag. So the content of the container should be much more predictable.

This also adds full container names instead of shortnames.

Closes: #870