chore: fix copyright year based on last update
The copyright tool did not look into test folder.
I think we should update the year IF the copyright header is there.
Having said that, I don't think test files need the header at all.

Since we missed to do that in the past future copyright runs will update the year to 2023 (since this commit will land in 2023) :-(