fix(binding-mqtt): subscribe after unsubscribe
When subscribing to a topic after unsubscribing from it, the subscription
is not performed until the client publish the `connect` event. Since it
was already connected with the first subscription the `connect` event
is not fired again. This commit fixes this issue by verifying if the
client is connected before subscribing to a topic.

Fixes #942

Signed-off-by: reluc <relu.cri@gmail.com>